requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click-repl==0.1.0
click==6.6
prompt-toolkit==1.0.8     # via click-repl
ruamel.yaml==0.12.15
six==1.10.0
SQLAlchemy==1.1.3
typing==3.5.2.2
wcwidth==0.1.7            # via prompt-toolkit
